third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-dom-k2-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'k2' attribute of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.getAttribute('k2') is ".9"
PASS overComposite1.getAttribute('k2') is ".1"
PASS successfullyParsed is true

TEST COMPLETE
